copilotkit-agent-py/pyproject.toml
[tool.poetry]
name = "tutorial_quickstart"
version = "0.1.0"
description = "Quickstart Tutorial"
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
license = "MIT"
packages = [{include = "sample_agent"}]

[project]
name = "tutorial_quickstart"
version = "0.0.1"
dependencies = [
  "langgraph==0.2.52",
  "langchain_core",
  "langchain-google-genai",
  "langchain",
  "openai",
  "langchain-community",
  "langchain-anthropic",
  "tavily-python",
  "copilotkit==0.1.41",
  "python-dotenv",
  "uvicorn",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "3.13.3"
copilotkit = "0.1.41"
langgraph = "0.2.52"
langchain-core = "^0.3.12"
langchain-community = "^0.3.1"
langchain-anthropic = "0.2.3"
langchain = "0.3.4"
openai = "^1.52.1"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"

[tool.poetry.scripts]
demo = "tutorial_quickstart.demo:main"